requirements.txt
blinker==1.4
click==8.1.1
colorama==0.4.4
Flask==2.1.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.33
Werkzeug==2.1.0
WTForms==3.0.1
zipp==3.7.0
